requirements.txt

numpy==1.25.2
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.2
opencv-python==4.8.1.78
psycopg2-binary
ultralytics
apscheduler
torch
lap
beautifulsoup4
tensorflow
selenium
xvfbwrapper
cython-bbox
